requirements.txt
ninja
Flask
gunicorn
dlib
torch
numpy
Pillow
torchvision
matplotlib
